[executor] split execute_and_commit_chunk into separate functions
This commit splits the execute_and_commit_chunk into execute and commit functions so that
we can run them in pipeline fashion to increase throughput in state sync v2.

Closes: #9505
